+++ CHANGELOG.md
@@ -8,6 +8,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The intended audience of this file is for py42 consumers -- as such, changes that don't affect
 how a consumer would use the library (e.g. adding unit tests, updating documentation, etc) are not captured here.
 
+## Unreleased
+
+- `sdk.cases.create()` now raises `Py42CaseNameExistsError` when the case name already exists in the system.
+
+- `sdk.cases.create()` now raises `Py42DescriptionLimitExceededError` when the description is more than 250 charachters.
+
+- `sdk.cases.update()` now raises `Py42DescriptionLimitExceededError` when the description is more than 250 charachters.
+
+- `sdk.cases.file_events.add()` now raises `Py42CaseAlreadyHasEventError` when the same event is added to a case more than once.
+
+- `sdk.cases.file_events.add()` now raises `Py42UpdateClosedCaseError` when the event is added to a closed case.
+
+- `sdk.cases.file_events.delete()` now raises `Py42UpdateClosedCaseError` when the event is deleted from a closed case.
+
 ## 1.11.0 - 2021-01-20
 
 ### Changed